+++ src/core/op_misc.c
@@ -1,20 +1,11 @@
-/*
- *  Sarien AGI :: Copyright (C) 1999 Dark Fiber 
- *
+/*  Sarien - A Sierra AGI resource interpreter engine
+ *  Copyright (C) 1999-2001 Stuart George and Claudio Matsuoka
+ *  
+ *  $Id$
  *
  *  This program is free software; you can redistribute it and/or modify
  *  it under the terms of the GNU General Public License as published by
- *  the Free Software Foundation; either version 2 of the License, or
- *  (at your option) any later version.
- *
- *  This program is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *  GNU General Public License for more details.
- *
- *  You should have received a copy of the GNU General Public License
- *  along with this program; if not, write to the Free Software
- *  Foundation, Inc., 675 Mass Ave, Cambridge, MA 02139, USA.
+ *  the Free Software Foundation; see docs/COPYING for further details.
  */
 
 #include <stdio.h>
@@ -35,13 +26,11 @@
 #define code (logics[lognum].data)
 
 
-//extern    UINT8   *gid;
-
 extern struct agi_view_table view_table[];
 extern struct agi_logic logics[];
 
 
-void new_room (UINT16 r)
+void new_room (int r)
 {
 	_D (("(%d)", r));
 
@@ -59,8 +48,7 @@ void new_room (UINT16 r)
 	erase_sprites ();
 	new_room_resources ();
 
-	switch (getvar (V_border_touch_ego))
-	{
+	switch (getvar (V_border_touch_ego)) {
 	case 1:
 		view_table[EGO_VIEW_TABLE].y_pos = _HEIGHT - 1;
 		break;
